+++ src/main/kotlin/no/nowak/gpstracker/core/password/PasswordRepository.kt
@@ -2,5 +2,5 @@ package no.nowak.gpstracker.core.password
 
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface PasswordRepository : JpaRepository<Password, Long>{
+interface PasswordRepository : JpaRepository<Password, Int>{
 }
